fixes cyclic dependency
